Switch to .diff for core PRs
`git am` works more reliably for patches on a git clone, especially containing multiple commits, and I must have only tested it on the patch applied from an extension PR, since that was where the original problem came up. But because we're using --prefer-dist to download core, `git am` doesn't work. However `git apply` works on any folder, but it's only reliable on the final diff file not a "mail patch".