Ignore missing email when coming from webhook
A few releases ago gitlab made a change where emails that aren't marked public now get sent to webhooks as the literal string "[REDACTED]". This action step is used though when e.g. running scheduled where there is no webhook, and so it's useful to notify _somebody_ on failure, which in that case comes from the email provided during CiviCARROT signup.